+++ designer/client/actions/nk/assignSettings.ts
@@ -8,11 +8,6 @@ export type MetricsType = {
   scenarioTypeToDashboard: UnknownRecord,
 }
 
-export type UsageStatisticsReports = {
-  enabled: boolean,
-  url: string,
-}
-
 export type FeaturesSettings = {
   counts: boolean,
   search: { url: string },
@@ -25,7 +20,6 @@ export type FeaturesSettings = {
   tabs: DynamicTabData[],
   testDataSettings?: TestDataSettings,
   redirectAfterArchive: boolean,
-  usageStatisticsReports: UsageStatisticsReports,
 };
 
 export type TestDataSettings = {
