+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/config/FeatureTogglesConfig.scala
@@ -20,7 +20,6 @@ case class FeatureTogglesConfig(development: Boolean,
                                 testDataSettings: TestDataSettings,
                                 enableConfigEndpoint: Boolean,
                                 redirectAfterArchive: Boolean,
-                                usageStatisticsReports: UsageStatisticsReportsConfig
                                )
 
 object FeatureTogglesConfig extends LazyLogging{
@@ -45,7 +44,6 @@ object FeatureTogglesConfig extends LazyLogging{
     val intervalTimeSettings = config.as[IntervalTimeSettings]("intervalTimeSettings")
     val testDataSettings = config.as[TestDataSettings]("testDataSettings")
     val redirectAfterArchive = config.getAs[Boolean]("redirectAfterArchive").getOrElse(true)
-    val usageStatisticsReports = config.as[UsageStatisticsReportsConfig]("usageStatisticsReports")
 
     FeatureTogglesConfig(
       development = isDevelopmentMode,
@@ -60,7 +58,6 @@ object FeatureTogglesConfig extends LazyLogging{
       testDataSettings = testDataSettings,
       enableConfigEndpoint = enableConfigEndpoint,
       redirectAfterArchive = redirectAfterArchive,
-      usageStatisticsReports = usageStatisticsReports
     )
   }
